Move reusable test helpers to testutil package
